Refactor PathoscopeBowtie for use with virtool.job package
